Create flattentree.java
Flatten Binary Tree To Linked List. Write a program that flattens a given binary tree to a linked list.

The sequence of nodes in the linked list should be the same as that of the preorder traversal of the binary tree.
The linked list nodes are the same binary tree nodes. You are not allowed to create extra nodes.
The right child of a node points to the next node of the linked list whereas the left child points to NULL.


We will use the intuition behind morris’s traversal. In Morris Traversal we use the concept of a threaded binary tree, i.e If we set the right child of every node and the left child as NULL, our job will be done.

Approach:

The algorithm can be described as:

At a node(say cur) if there exists a left child, we will find the rightmost node in the left subtree(say prev).
We will set prev’s right child to cur’s right child,
We will then set cur’s right child to it’s left child.
We will then move cur to the next node by assigning cur it to its right child
We will stop the execution when cur points to NULL.