Huffman Encoding Problem
Huffman decoding is a technique that converts the encoded data into initial data. As we have seen in encoding, the Huffman tree is made for an input string and the characters are decoded based on their position in the tree. The decoding process is as follows:

Start traversing over the tree from the root node and search for the character.
If we move left in the binary tree, add 0 to the code.
If we move right in the binary tree, add 1 to the code.
The child node holds the input character. It is assigned the code formed by subsequent 0s and 1s. The time complexity of decoding a string is O(n), where n is the length of the string.